Done: Screenshot update, user-interface improvements, code testing
Fixed bug (High): If you do a spell check on the text, and submit form, the textbox collapses and text is not sent. Actually, I had put the AJAX loader image in the proofread html, and also was changing its CSS via jQuery which made it difficult for the functions to compare the strings.

Coming Up: Compatibility with other plugins test, translation updates

Most probably, I will release v1.6 today